Maven library package works with libgdx dependencices.
The library Manifest file do not contains any class path.
Usefull infromation and implementation entries added to the Manifest.
License and Readme file added to the packed jar.
Library resources cleaned-up. They are now copied to the "res/lib" folder in the final jar. Instructions added in the resource Readme file.

The packed jar still contains natives libgdx libraries for the desktop. They should be removed in the future and added to the desktop project.